pyproject.toml
[tool.poetry]
name = "github_iterator_replace"
version = "0.1.0"
description = ""
authors = ["jevgenix <jekka9803@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.57"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
